Integrate libsys4 into build
No bindings yet, just the build infrastructure (consisting of dune,
meson, make *and* a shell script).

The makefile/script are there because dune doesn't integrate very well
with foreign build systems. It's easier to just invoke meson and dune
sequentially and have dune assume that libsys4.a is already built.