Use Alice.Ain.Type.t for type analysis
This is a type which is created from `struct ain_type` objects by the
libsys4 bindings. Using this type during type analysis will make it
easier to work with type data stored in the .ain file (globals, etc.).

Jaf.type_specifier is now purely a description of the input syntax, and
should not be used after the type analysis phase except to check for
const/override specifiers.