Finish codegen implementation
The codegen pass is now more or less complete, but almost entirely
untested. This includes support for some stuff that wasn't implemented
in the alice-tools compiler, such as the full set of built-ins described
in the SDK documentation.

Next step is to implement a proper command line interface and testing
infrastructure.