Add -declarations-only option (-d)
When this argument is supplied on the command line, only top-level
declarations are processed and no code is compiled into the output .ain
file.

This is currently not useful for anything, but will be used to achieve
separate compilation in the future. E.g., to compile a project with two
files, a.jaf and b.jaf, the following sequence of invocations would be
performed,

    # generate declarations file for a.jaf
    sys4c -d -o a.decl.ain a.jaf
    # generate declarations file for b.jaf
    sys4c -d -o b.decl.ain b.jaf
    # link all declarations files
    sys4c -d -o declarations.ain a.decl.ain b.decl.ain
    # compile a.jaf (importing declarations from declarations.ain)
    sys4c -c -o a.ain -i declarations.ain a.jaf
    # compile b.jaf (importing declarations from declarations.ain)
    sys4c -c -o b.ain -i declarations.ain b.jaf
    # link
    sys4c -o out.ain declarations.ain a.ain b.ain