Add -import-declarations (-i) option
This option allows specifying an ain file to import missing declarations
from (to be used with files produced using the -d option). The effect is
similar to providing the declarations ain file as an input file, except
that declarations are only imported as needed.

The intended usage pattern is as follows,

    sys4c -d -o a.decl.ain a.jaf
    sys4c -d -o b.decl.ain b.jaf
    sys4c -d -o declarations.jaf a.decl.jaf b.decl.jaf
    sys4c -c -i declarations.jaf -o a.ain a.jaf
    sys4c -c -i declarations.jaf -o b.ain b.jaf
    sys4c -o out.ain declarations.jaf a.ain b.ain