Add -compile-unit (-c) option
If the -c option is NOT specified, then the compiler will check for the
existence of undefined functions and throw an error before writing the
output file if there are any. The -c option simply disables this check.

Some linker bugs are also fixed here (function address not being
updated, main/message function index not being updated).

With this, it is now possible to implement a build system with separate
compilation. However, it would be necessary in such a system to specify
all dependencies manually. The final step will be to implement an option
which can check if two .ain files contain incompatible definitions.