Call destructors on goto/break/continue
Call destructors for any local variables that will go out of scope after
a jump.

This is implemented in the variable allocation pass. Each label, loop
and control flow statement is tagged with a set of variables that are
live at the beginning of that statement. The set of variables to be
deleted is obtained by taking the difference of the live variables at
the jump and target statements.